Build 1.05.01

+++ SocialsNetwork/Endpoints/Class/FriendRequests/FriendRequestPost.cs
@@ -1,4 +1,5 @@
-﻿using SocialsNetwork.DTO.Class;
+﻿using Azure.Core;
+using SocialsNetwork.DTO.Class;
 using SocialsNetwork.Endpoints.DicionaryErrors;
 using SocialsNetwork.Infra.Data;
 using SocialsNetwork.Models.Class;
@@ -19,14 +20,14 @@ namespace SocialsNetwork.Endpoints.Class.FriendRequests
            
             if (request.status.Equals(RequestStatus.accepted))
             {
-                // inserir novo registro na tabela friendships
+                // inserir novo registro na tabela friendships e apaga o registro da tabela FriendRequest
                 return await AddNewFriendship(LoggedUser, request, context);
-            } else if (request.status.Equals(RequestStatus.denied)
-                        || request.status.Equals(RequestStatus.canceled))
+            } else if (request.status.Equals(RequestStatus.denied))
             {
                 // apagar a solicitação de amizade de FriendRequest
-               return await RemoveFriendshipRequest(request.Id, context);   
+                return await RemoveFriendshipRequest(LoggedUser, request, context);   
             }
+
 
             return Results.Ok();
         }
@@ -46,7 +47,7 @@ namespace SocialsNetwork.Endpoints.Class.FriendRequests
 
             /* Verifica se o usuario que recebeu a solicitação é o mesmo que o logado */
             if (!IsRequested(asked.Id, request.Asked))
-                return Results.BadRequest();
+                return Results.Forbid();
 
             /* Cria a instancia de uma nova amizade e realiza a validação usando os dados de usuario
              * obtidos anteriormente 
@@ -56,21 +57,29 @@ namespace SocialsNetwork.Endpoints.Class.FriendRequests
                 return Results.ValidationProblem(data.Notifications.convertToDetails());
 
             await context.Friendships.AddAsync(data);
-            await RemoveFriendshipRequest(request.Id, context);
+            await RemoveFriendshipRequest(LoggedUser, request, context);
             await context.SaveChangesAsync();
-
             return Results.Created($"/Friend/{data.Id}", data.Id);
         }
 
-        private static async Task<IResult> RemoveFriendshipRequest(Guid IdRequest, AppDbContext context)
+        private static async Task<IResult> RemoveFriendshipRequest(string LoggedUser, FriendInviteManager request, AppDbContext context)
         {
-            var remove = context.FriendRequests.FindAsync(IdRequest).Result;
+           
+            var asked = context.ApplicationUsers.FindAsync(LoggedUser).Result; 
+            var askFriendship = context.ApplicationUsers.FindAsync(request.AskFriendship).Result;
+            
+            if (asked == null || askFriendship == null)
+                return Results.NotFound("Usuario não identificado");
 
-            if (remove == null) 
-                return Results.NotFound();
+            var remove = context.FriendRequests.FindAsync(request.Id).Result;
+            if (remove == null) return Results.NotFound("Requisição não identificada");
+
+            if (remove.Asked == null || remove.Asked.Id != asked.Id) return Results.Forbid();
 
             context.FriendRequests.Remove(remove);
-            await context.SaveChangesAsync();
+            /* Verifica se o status da requição é negado para casos onde o usuario negou a solicitação de amizade e salva */
+            if(request.status.Equals(RequestStatus.denied)) await context.SaveChangesAsync();
+
             return Results.Ok();
         }
         private static bool IsRequested(string Requested, string requestId)
